# Conflicts:
#	android/.gradle/6.9/executionHistory/executionHistory.bin
#	android/.gradle/6.9/executionHistory/executionHistory.lock
#	android/.gradle/6.9/fileHashes/fileHashes.bin
#	android/.gradle/6.9/fileHashes/fileHashes.lock
#	android/.gradle/6.9/fileHashes/resourceHashesCache.bin
#	android/.gradle/6.9/javaCompile/classAnalysis.bin
#	android/.gradle/6.9/javaCompile/jarAnalysis.bin
#	android/.gradle/6.9/javaCompile/javaCompile.lock
#	android/.gradle/6.9/javaCompile/taskHistory.bin
#	android/.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	android/.gradle/buildOutputCleanup/outputFiles.bin
#	android/.gradle/checksums/checksums.lock
#	android/.gradle/checksums/md5-checksums.bin
#	android/.gradle/checksums/sha1-checksums.bin
#	ios/Runner/Assets.xcassets/LaunchImage.imageset/Contents.json
#	lib/pages/leaderboard/leaderboard_screen.dart
#	lib/pages/profile/profile_screen.dart

+++ lib/services/auth/openday_login_service.dart
@@ -4,7 +4,11 @@ import 'dart:io';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:iscte_spots/models/auth/login_form_result.dart';
+import 'package:iscte_spots/models/database/tables/database_puzzle_piece_table.dart';
+import 'package:iscte_spots/pages/auth/auth_page.dart';
 import 'package:iscte_spots/services/auth/auth_service.dart';
+import 'package:iscte_spots/services/onboard_service.dart';
+import 'package:iscte_spots/services/shared_prefs_service.dart';
 import 'package:iscte_spots/widgets/util/constants.dart';
 import 'package:logger/logger.dart';
 
@@ -37,7 +41,7 @@ class OpenDayLoginService {
       );
     } else {
       _logger.e(
-          "statusCode: ${response.statusCode} on login response: $decodedResponse");
+          "statusCode: ${response.statusCode} with login response: $decodedResponse");
     }
     return response.statusCode;
   }
@@ -63,7 +67,41 @@ class OpenDayLoginService {
   }
 
   static Future<void> logOut(BuildContext context) async {
-    AuthService.deleteUserCredentials();
+    _logger.d("Logging Out User:");
+    /*const FlutterSecureStorage secureStorage = FlutterSecureStorage();
+
+    HttpClient client = HttpClient();
+    client.badCertificateCallback =
+        ((X509Certificate cert, String host, int port) => true);
+
+    final HttpClientRequest request = await client
+        .postUrl(Uri.parse('${BackEndConstants.API_ADDRESS}/api/auth/logout'));
+
+    String? apiToken = await secureStorage.read(key: "backend_api_key");
+    request.headers.add("Authorization", "Token $apiToken");
+
+    request.headers.set('content-type', 'application/json');
+
+    HttpClientResponse response = await request.close();
+    var decodedResponse =
+        await jsonDecode(await response.transform(utf8.decoder).join());
+    if (response.statusCode == 200) {
+*/
+    await AuthService.deleteUserCredentials();
+    await OnboadingService.removeOnboard();
+    Navigator.of(context).popUntil(ModalRoute.withName(AuthPage.pageRoute));
+    Navigator.of(context).pushNamed(AuthPage.pageRoute);
+    /*  } else {
+      _logger.e(
+          "statusCode: ${response.statusCode} on login response: $decodedResponse");
+    }
+*/
+    //Resetting status of completing all puzzles
+    SharedPrefsService.resetCompletedAllPuzzles();
+
+    //Removing placed puzzle pieces
+    DatabasePuzzlePieceTable.removeALL();
+
     return;
   }
 }
